+++ js/posts/thumbnails.js
